moving provider functions for item, missile, npc, player into package export file d2mapentity.go
